+++ spring-webmvc/src/test/java/org/springframework/web/servlet/mvc/method/RequestMappingInfoTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2019 the original author or authors.
+ * Copyright 2002-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -230,7 +230,7 @@ public class RequestMappingInfoTests {
 				.build();
 
 		assertThat(info1.equals(info2)).isFalse();
-		assertThat(info2.hashCode()).isNotEqualTo((long) info1.hashCode());
+		assertThat(info2.hashCode()).isNotEqualTo(info1.hashCode());
 
 		info2 = paths("/foo").methods(GET, RequestMethod.POST)
 				.params("foo=bar", "customFoo=customBar").headers("foo=bar")
@@ -238,7 +238,7 @@ public class RequestMappingInfoTests {
 				.build();
 
 		assertThat(info1.equals(info2)).isFalse();
-		assertThat(info2.hashCode()).isNotEqualTo((long) info1.hashCode());
+		assertThat(info2.hashCode()).isNotEqualTo(info1.hashCode());
 
 		info2 = paths("/foo").methods(GET)
 				.params("/NOOOOOO", "customFoo=customBar").headers("foo=bar")
@@ -246,7 +246,7 @@ public class RequestMappingInfoTests {
 				.build();
 
 		assertThat(info1.equals(info2)).isFalse();
-		assertThat(info2.hashCode()).isNotEqualTo((long) info1.hashCode());
+		assertThat(info2.hashCode()).isNotEqualTo(info1.hashCode());
 
 		info2 = paths("/foo").methods(GET)
 				.params("foo=bar", "customFoo=customBar").headers("/NOOOOOO")
@@ -254,7 +254,7 @@ public class RequestMappingInfoTests {
 				.build();
 
 		assertThat(info1.equals(info2)).isFalse();
-		assertThat(info2.hashCode()).isNotEqualTo((long) info1.hashCode());
+		assertThat(info2.hashCode()).isNotEqualTo(info1.hashCode());
 
 		info2 = paths("/foo").methods(GET)
 				.params("foo=bar", "customFoo=customBar").headers("foo=bar")
@@ -262,7 +262,7 @@ public class RequestMappingInfoTests {
 				.build();
 
 		assertThat(info1.equals(info2)).isFalse();
-		assertThat(info2.hashCode()).isNotEqualTo((long) info1.hashCode());
+		assertThat(info2.hashCode()).isNotEqualTo(info1.hashCode());
 
 		info2 = paths("/foo").methods(GET)
 				.params("foo=bar", "customFoo=customBar").headers("foo=bar")
@@ -270,7 +270,7 @@ public class RequestMappingInfoTests {
 				.build();
 
 		assertThat(info1.equals(info2)).isFalse();
-		assertThat(info2.hashCode()).isNotEqualTo((long) info1.hashCode());
+		assertThat(info2.hashCode()).isNotEqualTo(info1.hashCode());
 
 		info2 = paths("/foo").methods(GET)
 				.params("foo=bar", "customFoo=NOOOOOO").headers("foo=bar")
@@ -278,7 +278,7 @@ public class RequestMappingInfoTests {
 				.build();
 
 		assertThat(info1.equals(info2)).isFalse();
-		assertThat(info2.hashCode()).isNotEqualTo((long) info1.hashCode());
+		assertThat(info2.hashCode()).isNotEqualTo(info1.hashCode());
 	}
 
 	@Test
